Added NTSTATUS logging improvements
If the data printer detects an NTSTATUS error (identified by the upper 4 bytes being 0xFFFFFFFF [as a result from sign-extending], and the high nibble of the lower DWORD being 0xC) it will use FormatMessageA to produce a texture error message. Ideally this would just print out the definition name, but either way this makes it a lot easier to see what was logged.